Began move of db routes to database.js. One working route: /db/{collection}
 Changes to be committed:
	new file:   config/routes/database.js
	modified:   config/routes/index.js
	modified:   index.js
