+++ config/routes/index.js
@@ -1,3 +1,4 @@
 var includes = require('./includes');
+var database = require('./database');
 
-module.exports = [].concat(includes);
+module.exports = [].concat(includes,database);
